fix bug in check_environment.ipynb
In case the user doesn't have a required module, the `version`-list is shorter than the lists `packages` and `suggested_v`. The script prints
    [FAIL]: sklearn is not installed
    [OK] numpy 1.12.0
    [OK] scipy 0.18.1
    [OK] matplotlib 2.0.0
    [OK] sklearn 0.19.2
if the user doesn't have sklearn installed. Note that there is no print for pandas because the `version`-list is too short. The sklearn in the last line is wrong and should read pandas.